add new file update class

+++ Shape.cs
@@ -8,6 +8,6 @@ namespace OOpShapes
 {
     abstract class Shape
     {
-        public abstract int GetArea();
+        public abstract double GetArea();
     }
 }
